env: Redisson 의존성 추가 및 Redis 환경설정
