+++ gradle.properties
@@ -15,6 +15,6 @@ ktlintVersion=12.1.1
 
 ### External dependency versions ###
 springMockkVersion=4.0.2
-
-### JWT
-jwtVersion=0.11.2
+springdocOpenApiVersion=2.8.0
+redissonVersion=3.37.0
+tsidCreatorVersion=5.2.6
