+++ src/test/resources/application.yml
@@ -1,31 +1,4 @@
 spring:
-  application:
-    name: hhplus
-  profiles:
-    active: local
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    name: HangHaePlusDataSource
-    type: com.zaxxer.hikari.HikariDataSource
-    hikari:
-      pool-name: core-db-pool
-      maximum-pool-size: 3
-      connection-timeout: 10000
-      max-lifetime: 60000
-  jpa:
-    open-in-view: false
-    generate-ddl: false
-    show-sql: false
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate.timezone.default_storage: NORMALIZE_UTC
-      hibernate.jdbc.time_zone: UTC
-
----
-spring.config.activate.on-profile: test
-
-spring:
   jpa:
     show-sql: true
     hibernate:
